Test data generation for pubgrub impl
To test the pubgrub implementation test data is needed.
It would be nice to use the real world data available from
code.dlang.org.
But reading in the multiple MB json dumb for unit testing
is likely more work than actually running the tests with that data.
Or at least hopefully that is the case.

The functionality contained in inlinetestdatagen will read in
the dump from code.dlang.org once and then generate a D source
file that contains all packages spelled out as struct constructor
calls.
A poor mans binary format so to speak.